Remove unnecessary float cast in satpy/readers/abi_l1b.py

+++ satpy/readers/abi_l1b.py
@@ -133,7 +133,7 @@ class NC_ABI_L1B(NC_ABI_BASE):
     def _vis_calibrate(self, data):
         """Calibrate visible channels to reflectance."""
         solar_irradiance = self["esun"]
-        esd = self["earth_sun_distance_anomaly_in_AU"].astype(np.float32)
+        esd = self["earth_sun_distance_anomaly_in_AU"]
 
         factor = np.pi * esd * esd / solar_irradiance
